ai-interview-proctor/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
sqlalchemy
pydantic
opencv-python-headless
mediapipe
numpy
pillow
gTTS
python-dotenv
